+++ packages/sdks/web-component/src/lib/descope-wc/DescopeWc.ts
@@ -15,6 +15,7 @@ import {
   RESPONSE_ACTIONS,
   SDK_SCRIPTS_LOAD_TIMEOUT,
   URL_CODE_PARAM_NAME,
+  URL_ERR_PARAM_NAME,
   URL_RUN_IDS_PARAM_NAME,
   URL_TOKEN_PARAM_NAME,
 } from '../constants';
@@ -51,7 +52,8 @@ import {
   replaceElementMessage,
   setCssVars,
   setNOTPVariable,
-  setPhoneAutoDetectDefaultCode,
+  setComponentsAutoDetectByGeo,
+  setComponentsAutoDetectByLocale,
 } from '../helpers/templates';
 import {
   ClientScript,
@@ -180,12 +182,15 @@ class DescopeWc extends BaseDescopeWc {
     const response = JSON.parse(payload);
     if (type === 'oauthWeb' || type === 'sso') {
       let { exchangeCode } = response;
+      let exchangeError: string | undefined;
       if (!exchangeCode) {
         const url = new URL(response.url);
         exchangeCode = url.searchParams?.get(URL_CODE_PARAM_NAME);
+        exchangeError = url.searchParams?.get(URL_ERR_PARAM_NAME) || undefined;
       }
       this.nativeCallbacks.complete?.({
         exchangeCode,
+        exchangeError,
         idpInitiated: true,
       });
     } else if (type === 'magicLink') {
@@ -298,7 +303,7 @@ class DescopeWc extends BaseDescopeWc {
         }
         await this.injectNpmLib(
           '@descope/flow-scripts',
-          '1.0.13', // currently using a fixed version when loading scripts
+          '1.0.14', // currently using a fixed version when loading scripts
           `dist/${script.id}.js`,
         );
         const module = globalThis.descope?.[script.id];
@@ -1078,6 +1083,7 @@ class DescopeWc extends BaseDescopeWc {
       oidcErrorRedirectUri,
       oidcResource,
       action,
+      locale: getUserLocale(locale).locale,
     };
 
     const lastAuth = getLastAuth(loginId);
@@ -1620,9 +1626,10 @@ class DescopeWc extends BaseDescopeWc {
       this.loggerWrapper,
     );
 
-    // set the default country code based on the locale value we got
-    const { geo } = await this.getExecutionContext();
-    setPhoneAutoDetectDefaultCode(clone, geo);
+    // set auto-detect attributes (country code from geo, lang from locale)
+    const { geo } = (await this.getExecutionContext()) ?? {};
+    setComponentsAutoDetectByGeo(clone, geo);
+    setComponentsAutoDetectByLocale(clone, currentState.locale);
 
     const injectNextPage = async () => {
       await loadDescopeUiComponents;