+++ packages/widgets/applications-portal-widget/CHANGELOG.md
@@ -2,6 +2,43 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.6.0](https://github.com/descope/descope-js/compare/applications-portal-widget-0.5.22...applications-portal-widget-0.6.0) (2026-04-20)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.47.2`
+
+### Features
+
+* add WS-Fed apps to applications portal widget ([#1366](https://github.com/descope/descope-js/issues/1366)) ([22d8813](https://github.com/descope/descope-js/commit/22d8813655dbeaf213d1bb6bf018162481b92ebb))
+
+## [0.5.22](https://github.com/descope/descope-js/compare/applications-portal-widget-0.5.21...applications-portal-widget-0.5.22) (2026-04-13)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.47.1`
+## [0.5.21](https://github.com/descope/descope-js/compare/applications-portal-widget-0.5.20...applications-portal-widget-0.5.21) (2026-04-12)
+
+### Dependency Updates
+
+* `sdk-helpers` updated to version `0.7.0`
+* `sdk-mixins` updated to version `0.16.1`
+* `sdk-component-drivers` updated to version `0.9.1`
+## [0.5.20](https://github.com/descope/descope-js/compare/applications-portal-widget-0.5.19...applications-portal-widget-0.5.20) (2026-03-26)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.47.0`
+## [0.5.19](https://github.com/descope/descope-js/compare/applications-portal-widget-0.5.18...applications-portal-widget-0.5.19) (2026-03-23)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.46.0`
+## [0.5.18](https://github.com/descope/descope-js/compare/applications-portal-widget-0.5.17...applications-portal-widget-0.5.18) (2026-02-26)
+
+### Dependency Updates
+
+* `sdk-mixins` updated to version `0.16.0`
 ## [0.5.17](https://github.com/descope/descope-js/compare/applications-portal-widget-0.5.16...applications-portal-widget-0.5.17) (2026-01-29)
 
 ### Dependency Updates