+++ packages/widgets/tenant-profile-widget/CHANGELOG.md
@@ -2,6 +2,62 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.5.30](https://github.com/descope/descope-js/compare/tenant-profile-widget-0.5.29...tenant-profile-widget-0.5.30) (2026-04-20)
+
+### Dependency Updates
+
+* `core-js-sdk` updated to version `2.58.2`
+* `web-js-sdk` updated to version `1.47.2`
+* `web-component` updated to version `3.58.4`
+## [0.5.29](https://github.com/descope/descope-js/compare/tenant-profile-widget-0.5.28...tenant-profile-widget-0.5.29) (2026-04-13)
+
+### Dependency Updates
+
+* `core-js-sdk` updated to version `2.58.1`
+* `web-js-sdk` updated to version `1.47.1`
+* `web-component` updated to version `3.58.3`
+## [0.5.28](https://github.com/descope/descope-js/compare/tenant-profile-widget-0.5.27...tenant-profile-widget-0.5.28) (2026-04-12)
+
+### Dependency Updates
+
+* `web-component` updated to version `3.58.2`
+* `sdk-helpers` updated to version `0.7.0`
+* `sdk-mixins` updated to version `0.16.1`
+* `sdk-component-drivers` updated to version `0.9.1`
+## [0.5.27](https://github.com/descope/descope-js/compare/tenant-profile-widget-0.5.26...tenant-profile-widget-0.5.27) (2026-04-06)
+
+### Dependency Updates
+
+* `web-component` updated to version `3.58.1`
+## [0.5.26](https://github.com/descope/descope-js/compare/tenant-profile-widget-0.5.25...tenant-profile-widget-0.5.26) (2026-03-31)
+
+### Dependency Updates
+
+* `web-component` updated to version `3.58.0`
+## [0.5.25](https://github.com/descope/descope-js/compare/tenant-profile-widget-0.5.24...tenant-profile-widget-0.5.25) (2026-03-26)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.47.0`
+* `web-component` updated to version `3.57.1`
+## [0.5.24](https://github.com/descope/descope-js/compare/tenant-profile-widget-0.5.23...tenant-profile-widget-0.5.24) (2026-03-23)
+
+### Dependency Updates
+
+* `core-js-sdk` updated to version `2.58.0`
+* `web-js-sdk` updated to version `1.46.0`
+* `web-component` updated to version `3.57.0`
+## [0.5.23](https://github.com/descope/descope-js/compare/tenant-profile-widget-0.5.22...tenant-profile-widget-0.5.23) (2026-02-26)
+
+### Dependency Updates
+
+* `web-component` updated to version `3.56.0`
+* `sdk-mixins` updated to version `0.16.0`
+## [0.5.22](https://github.com/descope/descope-js/compare/tenant-profile-widget-0.5.21...tenant-profile-widget-0.5.22) (2026-02-26)
+
+### Dependency Updates
+
+* `web-component` updated to version `3.55.4`
 ## [0.5.21](https://github.com/descope/descope-js/compare/tenant-profile-widget-0.5.20...tenant-profile-widget-0.5.21) (2026-02-17)
 
 ### Dependency Updates